Issue 3903 - fix repl keep alive event interval
Description:  Previously we passed the interval as seconds to the
              event queue, but it is supposed to be milliseconds.

              Fixed a crash with repl logging and decoding extended
              op payload (referrals).

              Also reworked alot of the replication CI tests that
              were flaky.

relates: https://github.com/389ds/389-ds-base/issues/3903

Reviewed by: tbordaz & spichugi(Thanks!)
